+++ apps/web/src/test/setup.ts
@@ -17,3 +17,18 @@ Object.defineProperty(globalThis, "PublicKeyCredential", {
   configurable: true,
   value: class PublicKeyCredential {},
 });
+
+Object.defineProperty(HTMLDialogElement.prototype, "showModal", {
+  configurable: true,
+  value(this: HTMLDialogElement) {
+    this.setAttribute("open", "");
+  },
+});
+
+Object.defineProperty(HTMLDialogElement.prototype, "close", {
+  configurable: true,
+  value(this: HTMLDialogElement) {
+    this.removeAttribute("open");
+    this.dispatchEvent(new Event("close"));
+  },
+});